fix authentication filter to check the user
before prehandle->getUserFromSession->findById was not finding
any user. now it will check all users for the id instead of using
findById. no idea why findById is not working since the id looks right.
at least now the filtering works, if the session is valid the
search api will work with postman and if not it will fail with UNAUTHORIZED. login/register
work too and axios seems to use the cookie so no code changes needed to
use it.

frontend still doesn't work as userdetails is null and its needed for
a lot of the pages to work fully.